[Util] function to get port number
Add util function to get available port.

TODO: update testcase (data transfer with fixed port) later.
